+++ de-wordpress-plugin-utils.php
@@ -14,7 +14,7 @@ if ( defined( 'DE_WORDPRESS_PLUGIN_UTILS_VERSION' ) ) {
 	return;
 }
 
-define( 'DE_WORDPRESS_PLUGIN_UTILS_VERSION', '1.6.0' );
+define( 'DE_WORDPRESS_PLUGIN_UTILS_VERSION', '1.7.0' );
 define( 'DE_WORDPRESS_PLUGIN_UTILS_PATH', __DIR__ );
 
 require_once DE_WORDPRESS_PLUGIN_UTILS_PATH . '/src/Common/class-functions.php';
@@ -34,3 +34,5 @@ require_once DE_WORDPRESS_PLUGIN_UTILS_PATH . '/src/Abstracts/class-abstract-bac
 require_once DE_WORDPRESS_PLUGIN_UTILS_PATH . '/src/RestApi/class-rest-authentication.php';
 require_once DE_WORDPRESS_PLUGIN_UTILS_PATH . '/src/RestApi/class-rest-response.php';
 require_once DE_WORDPRESS_PLUGIN_UTILS_PATH . '/src/Settings/class-plugin-settings-repository.php';
+require_once DE_WORDPRESS_PLUGIN_UTILS_PATH . '/src/Export/class-export-csv.php';
+require_once DE_WORDPRESS_PLUGIN_UTILS_PATH . '/src/Upload/class-csv-upload.php';