+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.7.0] - 2026-04-28
+* ENH: Add CSV Export & Upload classes.
+* ENH: Add mysql_datetime_to_datetime_short function.
+
 ## [1.6.0] - 2026-04-26
 * ENH: Make Plugin_Settings_Repository more robust.